hack to make it work on 64bit machines.
The original code was developed on Intel Core2-Quad platform as in the original Kendo paper
because the platform has a known deterministic hardware counter (store counts). At the moment,
it is now clear whether such a deterministic counter exsits on more recent platforms.

Also, the code was originally developed in 32bit mode. Therefore running it on 64bit machines is
not tested.